move MultiChannelParameter to parameters module
