+++ qcodes/instrument_drivers/QDev/QDac_channels.py
@@ -9,8 +9,7 @@ import pyvisa as visa
 from pyvisa.resources.serial import SerialInstrument
 
 from qcodes.instrument import ChannelList, Instrument, InstrumentChannel, VisaInstrument
-from qcodes.instrument.channel import MultiChannelInstrumentParameter
-from qcodes.parameters import ParamRawDataType
+from qcodes.parameters import MultiChannelInstrumentParameter, ParamRawDataType
 from qcodes.utils import validators as vals
 
 log = logging.getLogger(__name__)
